// featurehub-javascript-client-sdk/app/models/models/rollout-strategy-field-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FeatureServiceApi
 * This describes the API clients use for accessing features
 *
 * The version of the OpenAPI document: 1.1.8
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */
export enum RolloutStrategyFieldType {
    String = 'STRING',
    SemanticVersion = 'SEMANTIC_VERSION',
    Number = 'NUMBER',
    Date = 'DATE',
    Datetime = 'DATETIME',
    Boolean = 'BOOLEAN',
    IpAddress = 'IP_ADDRESS'
}